# perf: Don't import Panel and Pandas directly in IPython