--- a/holoviews/ipython/__init__.py
+++ b/holoviews/ipython/__init__.py
@@ -29,7 +29,7 @@ def __getattr__(attr):
     if attr == "IPTestCase":
         from ..element.comparison import IPTestCase
         from ..util.warnings import deprecated
-        deprecated("1.22.0", old="holoviews.ipython.IPTestCase", new="holoviews.element.comparison.IPTestCase")
+        deprecated("1.23.0", old="holoviews.ipython.IPTestCase", new="holoviews.element.comparison.IPTestCase")
         return IPTestCase
     raise AttributeError(f"module {__name__} has no attribute {attr}")