--- a/holoviews/ipython/__init__.py
+++ b/holoviews/ipython/__init__.py
@@ -31,7 +31,7 @@ def __getattr__(attr):
         from ..util.warnings import deprecated
         deprecated("1.23.0", old="holoviews.ipython.IPTestCase", new="holoviews.element.comparison.IPTestCase")
         return IPTestCase
-    raise AttributeError(f"module {__name__} has no attribute {attr}")
+    raise AttributeError(f"module {__name__!r} has no attribute {attr!r}")
 
 
 class notebook_extension(extension):