--- a/holoviews/tests/util/test_init.py
+++ b/holoviews/tests/util/test_init.py
@@ -2,6 +2,8 @@ import sys
 from subprocess import check_output
 from textwrap import dedent
 
+import pytest
+
 
 def test_no_blocklist_imports():
     check = """\
@@ -13,8 +15,25 @@ def test_no_blocklist_imports():
 
     if mods:
         print(", ".join(mods), end="")
-        """
+    """
 
     output = check_output([sys.executable, '-c', dedent(check)])
+    assert output == b""
 
+
+def test_no_blocklist_imports_IPython():
+    pytest.importorskip("IPython")
+
+    check = """\
+    import sys
+    import holoviews as hv
+
+    blocklist = {"panel", "datashader", "ibis", "pandas"}
+    mods = blocklist & set(sys.modules)
+
+    if mods:
+        print(", ".join(mods), end="")
+    """
+
+    output = check_output([sys.executable, '-m', 'IPython', '-c', dedent(check)])
     assert output == b""