Continuing from https://github.com/holoviz/holoviews/pull/6476

I noticed that HoloViews was _very_ slow when importing in IPython. This is mainly because of the import of Panel and Pandas. This PR defers these imports. 

# Hyperfine

## Before (based on #6476)
 `hyperfine "ipython -c ''" "ipython -c 'import holoviews'" --warmup 5`

``` bash
❯ hyperfine "ipython -c ''" "ipython -c 'import holoviews'" --warmup 5
Benchmark 1: ipython -c ''
  Time (mean ± σ):     256.3 ms ±   2.9 ms    [User: 191.0 ms, System: 36.7 ms]
  Range (min … max):   252.2 ms … 260.9 ms    11 runs

Benchmark 2: ipython -c 'import holoviews'
  Time (mean ± σ):      1.398 s ±  0.017 s    [User: 2.820 s, System: 0.150 s]
  Range (min … max):    1.369 s …  1.416 s    10 runs

Summary
  ipython -c '' ran
    5.45 ± 0.09 times faster than ipython -c 'import holoviews'
```

## After (this PR)
``` bash
❯ hyperfine "ipython -c ''" "ipython -c 'import holoviews'" --warmup 5
Benchmark 1: ipython -c ''
  Time (mean ± σ):     256.7 ms ±   2.8 ms    [User: 194.2 ms, System: 34.5 ms]
  Range (min … max):   253.8 ms … 262.2 ms    11 runs

Benchmark 2: ipython -c 'import holoviews'
  Time (mean ± σ):     484.8 ms ±   5.7 ms    [User: 2020.4 ms, System: 57.8 ms]
  Range (min … max):   477.9 ms … 496.4 ms    10 runs

Summary
  ipython -c '' ran
    1.89 ± 0.03 times faster than ipython -c 'import holoviews'
```

# Tuna
`python -X importtime -m IPython -c 'import holoviews' 2> tuna.log && tuna tuna.log` (and not having setuptools_scm installed)

## Before (based on #6476)
![image](https://github.com/user-attachments/assets/0453a39d-0608-445c-a650-d26309fbe66c)

## After (this PR)
![image](https://github.com/user-attachments/assets/b590e704-f413-41fb-b36f-1fc3b41346ff)